+++ src/components/offer-list/offer-list.tsx
@@ -1,7 +1,7 @@
 ﻿import PlaceCard from '@/components/place-card/place-card.tsx';
-import {useAppDispatch} from '@/components/hooks/use-app-dispatch.tsx';
 import {setSelectedOffer} from '@/store/action.ts';
 import {Offer} from '@/types/api.ts';
+import {useAppDispatch} from '@/hooks/use-app-dispatch.tsx';
 
 export interface OfferListProps {
   offers: Offer[];
