+++ src/hooks/use-app-selector.tsx
