+++ src/pages/main-screen/main-screen.tsx
@@ -1,10 +1,10 @@
 ﻿import Header from '@/components/header/header.tsx';
 import CityList from '@/components/city-list/city-list.tsx';
-import {useAppSelector} from '@/components/hooks/use-app-selector.tsx';
 import MainOfferList from '@/components/main-offer-list/main-offer-list.tsx';
 import EmptyMainOfferList from '@/components/empty-main-offer-list/empty-main-offer-list.tsx';
 import Spinner from '@/components/spinner/spinner.tsx';
 import {cities} from '@/constants/cities.ts';
+import {useAppSelector} from '@/hooks/use-app-selector.tsx';
 
 export default function MainScreen(): JSX.Element {
   const city = useAppSelector((state) => state.city);
